Disable auto-commit for JDBC connections
Execution of SQL interfaces aborts with an error when using
the Postgres JDBC driver in default auto-commit mode. This
is because DbImpl::executeImpl always attempts to commit an
open connection, irrespective if a transaction is open.

The error returned by the Postgres database is:

    25P01 : no_active_sql_transaction

This patch disables auto-commit for all connections.